HttpTrigger/requirements.txt
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.30.1
azure-mgmt-core==1.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
